+++ example/.eslintrc.js
@@ -1,9 +1,0 @@
-module.exports = {
-  "extends": "airbnb",
-  "rules": {
-    "object-curly-newline": "off",
-    "import/prefer-default-export": "off",
-    "react/jsx-one-expression-per-line" : "off",
-    "react/prop-types" : "off"
-  }
-};