+++ example/src/Error/load.js
@@ -1,4 +1,7 @@
 import { formRegion } from '../shared/regionForm';
 import { fetchValueWithError } from '../shared/fetch';
 
-export const loadValueWithError = () => formRegion.load('valueWithError', fetchValueWithError);
+export const loadValueWithError = () => {
+  formRegion.load('valueWithError1', fetchValueWithError, { format: (result, snapshot, error) => (error ? null : result) });
+  formRegion.load('valueWithError2', fetchValueWithError, { format: (result, snapshot, error) => (error ? null : result) });
+};