+++ example/src/LoadingControl/components/Control3.jsx
@@ -1,13 +1,13 @@
 import React, { Fragment } from 'react';
 import { connectWith } from 'region-shortcut';
-import Loading from './Loading';
+import RawLoading from './Loading';
 import Display from './Display';
 
-const Loading3 = ({ user, follower }) => (
+const Loading = ({ user, follower }) => (
   <Fragment>
-    <Loading loading />
+    <RawLoading loading />
     <Display user={user} follower={follower} />
   </Fragment>
 );
 
-export default connectWith(['user', 'follower'], Display, Loading3);
+export default connectWith(['user', 'follower'], Display, { Loading });