+++ example/src/LoadingFull/index.jsx
@@ -4,11 +4,13 @@ import { Card } from 'antd';
 import { loadFollower } from '../shared/load';
 import Lines from '../shared/Lines';
 
+const extra = <a onClick={loadFollower}>More</a>; // eslint-disable-line
+
 const Display = ({ loading, user, follower }) => (
   <Card
     loading={loading}
     title={user}
-    extra={<a onClick={loadFollower}>More</a>}
+    extra={extra}
     style={{ width: 300, margin: 30 }}
   >
     <Lines lines={follower} />