+++ example/src/SideEffect/load.js
@@ -4,7 +4,6 @@ import { fetchFollower } from '../shared/fetch';
 set('sideEffect', null);
 
 export const loadFollowerWithSideEffect = () => load('follower', fetchFollower, {
-  forceUpdate: true,
   format: (result, snapshot = []) => {
     snapshot.push(result);
     set('sideEffect', snapshot.length);