+++ src/util/__test__/reducer.test.js
@@ -1,6 +1,6 @@
 import { region } from './region';
 
-const { reducer, private_actionTypes } = region;
+const { private_reducer, private_actionTypes } = region;
 
 const { SET } = private_actionTypes;
 
@@ -35,7 +35,7 @@ mockDate();
 describe('reducer', () => {
   test('set string', () => {
     const result = 'a user';
-    const state = reducer({}, { type: SET, payload: { key: 'user', result } });
+    const state = private_reducer({}, { type: SET, payload: { key: 'user', result } });
     expect(state).toEqual({
       errors: { user: undefined },
       fetchTimes: { user: 0 },
@@ -46,7 +46,7 @@ describe('reducer', () => {
 
   test('set array', () => {
     const result = [{ id: 1, name: 'zhangcong' }, { id: 2, name: 'milly' }];
-    const state = reducer({}, { type: SET, payload: { key: 'user', result } });
+    const state = private_reducer({}, { type: SET, payload: { key: 'user', result } });
     expect(state).toEqual({
       errors: { user: undefined },
       fetchTimes: { user: 0 },
@@ -57,12 +57,35 @@ describe('reducer', () => {
 
   test('function', () => {
     const result = () => 'should not be string';
-    const state = reducer({}, { type: SET, payload: { key: 'user', result } });
+    const state = private_reducer({}, { type: SET, payload: { key: 'user', result } });
     expect(state).toEqual({
       errors: { user: undefined },
       fetchTimes: { user: 0 },
       loadings: { user: 0 },
       results: { user: result },
+    });
+  });
+
+  test('error', () => {
+    const error = new Error('error');
+    const state = private_reducer({}, { type: SET, payload: { key: 'user', error } });
+    expect(state).toEqual({
+      errors: { user: error },
+      fetchTimes: { user: 0 },
+      loadings: { user: 0 },
+    });
+  });
+
+  test('error not cover snapshot', () => {
+    const result = 'a user';
+    const error = new Error('error');
+    const state = private_reducer({}, { type: SET, payload: { key: 'user', result } });
+    const stateWithError = private_reducer(state, { type: SET, payload: { key: 'user', error } });
+    expect(stateWithError).toEqual({
+      errors: { user: error },
+      fetchTimes: { user: 0 },
+      loadings: { user: 0 },
+      results: { user: 'a user' },
     });
   });
 });